requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
firebase-admin==6.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10
langgraph==0.0.20
openai==1.6.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
eventlet==0.33.3